requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
absl-py=1.1.0=pyhd8ed1ab_0
aiohttp=3.8.1=py39h63b48b0_1
aiosignal=1.2.0=pyhd8ed1ab_0
async-timeout=4.0.2=pyhd8ed1ab_0
attrs=21.4.0=pyhd8ed1ab_0
blas=1.0=mkl
blinker=1.4=py_1
brotlipy=0.7.0=py39h9ed2024_1003
bzip2=1.0.8=h1de35cc_0
c-ares=1.18.1=h0d85af4_0
ca-certificates=2022.6.15=h033912b_0
cachetools=5.0.0=pyhd8ed1ab_0
certifi=2022.6.15=py39h6e9494a_0
cffi=1.15.0=py39hc55c11b_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.3=py39h6e9494a_0
colorama=0.4.5=pyhd8ed1ab_0
cryptography=37.0.1=py39hf6deb26_0
dataclasses=0.8=pyhc8e2a94_3
docker-pycreds=0.4.0=py_0
ffmpeg=4.3=h0a44026_0
freetype=2.11.0=hd8bbffd_0
frozenlist=1.3.0=py39h63b48b0_1
fsspec=2022.5.0=pyhd8ed1ab_0
future=0.18.2=py39h6e9494a_5
gettext=0.21.0=h7535e17_0
giflib=5.2.1=haf1e3a3_0
gitdb=4.0.9=pyhd8ed1ab_0
gitpython=3.1.27=pyhd8ed1ab_0
gmp=6.2.1=he9d5cce_3
gnutls=3.6.15=hed9c0bf_0
google-auth=2.9.1=pyh6c4a22f_0
google-auth-oauthlib=0.4.6=pyhd8ed1ab_0
grpcio=1.41.1=py39h757451a_1
icu=58.2=h0a44026_3
idna=3.3=pyhd3eb1b0_0
importlib-metadata=4.11.4=py39h6e9494a_0
intel-openmp=2021.4.0=hecd8cb5_3538
jpeg=9e=hca72f7f_0
lame=3.100=h1de35cc_0
lcms2=2.12=hf1fd2bf_0
libcxx=12.0.0=h2f01273_0
libffi=3.3=hb1e8313_2
libiconv=1.16=hca72f7f_2
libidn2=2.3.2=h9ed2024_0
libpng=1.6.37=ha441bb4_0
libprotobuf=3.15.8=hcf210ce_1
libtasn1=4.16.0=h9ed2024_0
libtiff=4.2.0=hdb42f99_1
libunistring=0.9.10=h9ed2024_0
libwebp=1.2.2=h56c3ce4_0
libwebp-base=1.2.2=hca72f7f_0
libxml2=2.9.14=hbf8cd5e_0
libzlib=1.2.12=hfe4f2af_2
lightning-bolts=0.5.0=pyhd8ed1ab_0
llvm-openmp=12.0.0=h0dcd299_1
lz4-c=1.9.3=h23ab428_1
markdown=3.3.7=pyhd8ed1ab_0
mkl=2021.4.0=hecd8cb5_637
mkl-service=2.4.0=py39h9ed2024_0
mkl_fft=1.3.1=py39h4ab4a9b_0
mkl_random=1.2.2=py39hb2f4e1b_0
multidict=6.0.2=py39h63b48b0_1
ncurses=6.3=hca72f7f_3
nettle=3.7.3=h230ac6f_1
numpy=1.22.3=py39h2e5f0a9_0
numpy-base=1.22.3=py39h3b1a694_0
oauthlib=3.2.0=pyhd8ed1ab_0
openh264=2.1.1=h8346a28_0
openssl=1.1.1q=hfe4f2af_0
packaging=21.3=pyhd8ed1ab_0
pathtools=0.1.2=py_1
pillow=9.2.0=py39hde71d04_0
pip=22.1.2=py39hecd8cb5_0
promise=2.3=py39h6e9494a_6
protobuf=3.15.8=py39h9fcab8e_0
psutil=5.9.1=py39h701faf5_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.21=pyhd3eb1b0_0
pydeprecate=0.3.2=pyhd8ed1ab_0
pyjwt=2.4.0=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=pyhd8ed1ab_0
pysocks=1.7.1=py39hecd8cb5_0
python=3.9.12=hdfd78df_1
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytorch=1.12.0=py3.9_0
pytorch-lightning=1.6.4=pyhd8ed1ab_0
pyu2f=0.1.5=pyhd8ed1ab_0
pyyaml=6.0=py39h63b48b0_4
readline=8.1.2=hca72f7f_1
requests=2.28.0=py39hecd8cb5_0
requests-oauthlib=1.3.1=pyhd8ed1ab_0
rsa=4.8=pyhd8ed1ab_0
sentry-sdk=1.7.1=pyhd8ed1ab_0
setproctitle=1.2.3=py39h63b48b0_0
setuptools=59.5.0=py39h6e9494a_0
shortuuid=1.0.9=pyha770c72_1
six=1.16.0=pyhd3eb1b0_1
smmap=3.0.5=pyh44b312d_0
sqlite=3.38.5=h707629a_0
tensorboard=2.9.1=pyhd8ed1ab_0
tensorboard-data-server=0.6.0=py39h1644bb1_2
tensorboard-plugin-wit=1.8.1=pyhd8ed1ab_0
tk=8.6.12=h5d9f67b_0
torchaudio=0.12.0=py39_cpu
torchmetrics=0.9.2=pyhd8ed1ab_0
torchvision=0.13.0=py39_cpu
tqdm=4.64.0=pyhd8ed1ab_0
typing-extensions=4.1.1=hd3eb1b0_0
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2022a=hda174b7_0
urllib3=1.26.9=py39hecd8cb5_0
wandb=0.12.21=pyhd8ed1ab_0
werkzeug=2.1.2=pyhd8ed1ab_1
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=hca72f7f_1
yaml=0.2.5=h0d85af4_2
yarl=1.7.2=py39h63b48b0_2
zipp=3.8.0=pyhd8ed1ab_0
zlib=1.2.12=h4dc903c_2
zstd=1.5.2=hcb37349_0
